requirements.txt
colorama==0.4.6
exceptiongroup==1.1.1
iniconfig==2.0.0
lxml>=4.9.2
ooo-dev-tools>=0.11.8
oooenv>=0.2.1
ooouno>=2.1.2
packaging==23.1
pluggy==1.0.0
pytest>=7.3.2
tomli==2.0.1
types-scriptforge>=2.0.1
types-uno-script>=0.1.1
types-unopy>=1.2.3
typing_extensions==4.6.3
